move config_schema to submodule
